hagelskott-analys/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
python-dotenv==1.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.3
pillow==10.2.0
pydantic==2.5.3
aiofiles==23.2.1
pydantic-settings==2.1.0
python-magic==0.4.27
email-validator==2.1.0.post1
logging-formatter-anticrlf==1.2
python-json-logger==2.0.7
httpx==0.26.0
tenacity==8.2.3
psutil==5.9.8
aiohttp==3.9.1
dataclasses==0.6
typing-extensions>=4.11,<5
